Rename .feature package to .ui
