+++ app/src/main/java/dev/burnoo/demo/listapp/composable/AppRouter.kt
@@ -3,8 +3,8 @@ package dev.burnoo.demo.listapp.composable
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
-import dev.burnoo.demo.listapp.userlist.UserListRoutes
-import dev.burnoo.demo.listapp.userlist.userListRouter
+import dev.burnoo.demo.listapp.ui.userlist.UserListRoutes
+import dev.burnoo.demo.listapp.ui.userlist.userListRouter
 
 @Composable
 fun AppRouter() {
