+++ app/src/main/java/dev/burnoo/demo/listapp/di/WithDependencyInjection.kt
@@ -2,7 +2,7 @@ package dev.burnoo.demo.listapp.di
 
 import androidx.compose.runtime.Composable
 import dev.burnoo.cokoin.Koin
-import dev.burnoo.demo.listapp.userlist.di.uiUserListModule
+import dev.burnoo.demo.listapp.ui.userlist.di.uiUserListModule
 
 @Composable
 fun WithDependencyInjection(content: @Composable () -> Unit) {
