+++ ui/userlist/src/main/java/dev/burnoo/demo/listapp/ui/userlist/composables/UserListContent.kt
@@ -1,4 +1,4 @@
-package dev.burnoo.demo.listapp.userlist.composables
+package dev.burnoo.demo.listapp.ui.userlist.composables
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -6,8 +6,8 @@ import androidx.lifecycle.compose.ExperimentalLifecycleComposeApi
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import dev.burnoo.cokoin.viewmodel.getViewModel
 import dev.burnoo.demo.listapp.data.users.model.UserId
-import dev.burnoo.demo.listapp.userlist.UserListUiState
-import dev.burnoo.demo.listapp.userlist.UserListViewModel
+import dev.burnoo.demo.listapp.ui.userlist.UserListUiState
+import dev.burnoo.demo.listapp.ui.userlist.UserListViewModel
 
 @OptIn(ExperimentalLifecycleComposeApi::class)
 @Composable
