+++ ui/userlist/src/main/java/dev/burnoo/demo/listapp/ui/userlist/di/Module.kt
@@ -1,7 +1,7 @@
-package dev.burnoo.demo.listapp.userlist.di
+package dev.burnoo.demo.listapp.ui.userlist.di
 
 import dev.burnoo.demo.listapp.data.users.core.di.usersRepositoryModule
-import dev.burnoo.demo.listapp.userlist.UserListViewModel
+import dev.burnoo.demo.listapp.ui.userlist.UserListViewModel
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.module
