+++ ui/userlist/src/test/kotlin/dev/burnoo/demo/listapp/userlist/UserListViewModelTest.kt
@@ -2,6 +2,8 @@ package dev.burnoo.demo.listapp.userlist
 
 import dev.burnoo.demo.listapp.data.users.core.FakeUsersRepository
 import dev.burnoo.demo.listapp.data.users.core.testUsers
+import dev.burnoo.demo.listapp.ui.userlist.UserListUiState
+import dev.burnoo.demo.listapp.ui.userlist.UserListViewModel
 import io.kotest.matchers.shouldBe
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
